fix model field

+++ adminpage/accounts/models.py
@@ -41,7 +41,7 @@ class User(AbstractUser):
     username = None
     email = models.EmailField(_('email address'), unique=True)
     role = models.CharField("role", max_length=100, blank=True, null=True)
-    telegram_id = models.IntegerField(null=True, blank=True)
+    telegram_id = models.PositiveBigIntegerField(null=True, blank=True)
 
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS = []
